9x9 improvement
Rows 0-6 work fine. Because the system checks per line, and runs a new sequence of numbers each time a repeated number is spotted, the last lines have a higher chance of running the horizontal assignment longer in search of a new sequence. Lines 7&8 have to run quicker. Consider adding a new function or exception for those rows that would check all previous row values and assign the remaining unused values.