9x9 grid works
I thought it was taking longer to load the grid because of the last two rows, however the debuggers were taking up most of the time. Once I commented them out, it only took less than 3 seconds to load the randomized grid.

Now I will create another array that will store the numbers and the blank spaces. Additionally, I will add a switch in which the user will be required to choose their difficulty level. The difficulty level will determine the number of blank spots on the 9x9 grid.